Restore audio player
